refactor: 이전 코드리뷰 반영
