+++ src/app/main/_components/news-carousel.tsx
